Metadata (#12)
Using a hard-coded local database, metadata for each track was acquired
from the database and output to the tracks section. There are a couple
of lingering issues we will handle in other PRs:
- The database is not created via the app, and the only way to access
the database is if it is included in the build files, meaning that users
would have to generate the database, insert it into the app, and rebuild
the app themselves every time there is any kind of update to the music
files on their phone
- Obviously, we will switch to having the app create the database
on-device for the user
- Another requirement will be that the database should be able to be
created and updated without rebuilding the app
  - See: https://github.com/JeffreyGaydos/SynthiaMusicApp/issues/13
- Partially because of the above, the images for the music files are not
able to be properly loaded in this PR (#14)
- The performance to load 2000+ songs and all their metadata is
extremely slow compared to the instantaneous loading speed of Samsung
Music. This will need fixed (#15)